refactor: move catch_issues into issue module
